// 2018/qiwei/src/app/g-data-vis/models/lineResult.ts
// this class define the lineResult data structure

export class LineResult {
    edge_name: string;
    tap_bus_name: string;
    z_bus_name: string;
    PA: number;
    PB: number;
    PC: number;
    QA: number;
    QB: number;
    QC: number;
    IA: number;
    IB: number;
    IC: number;
    IA_ang: number;
    IB_ang: number;
    IC_ang: number;
    lossA: number;
    lossB: number;
    lossC: number;
    startnode: string;
}